+++ core/src/main/java/com/github/shadowsocks/database/Profile.kt
@@ -21,6 +21,7 @@
 package com.github.shadowsocks.database
 
 import android.net.Uri
+import android.os.Parcelable
 import android.util.Base64
 import android.util.Log
 import android.util.LongSparseArray
@@ -32,6 +33,7 @@ import com.github.shadowsocks.preference.DataStore
 import com.github.shadowsocks.utils.Key
 import com.github.shadowsocks.utils.asIterable
 import com.github.shadowsocks.utils.parsePort
+import kotlinx.android.parcel.Parcelize
 import org.json.JSONArray
 import org.json.JSONObject
 import org.json.JSONTokener
@@ -41,7 +43,31 @@ import java.net.URISyntaxException
 import java.util.*
 
 @Entity
-class Profile : Serializable {
+@Parcelize
+data class Profile(
+        @PrimaryKey(autoGenerate = true)
+        var id: Long = 0,
+        var name: String? = "",
+        var host: String = "198.199.101.152",
+        var remotePort: Int = 8388,
+        var password: String = "u1rRWTssNv0p",
+        var method: String = "aes-256-cfb",
+        var route: String = "all",
+        var remoteDns: String = "8.8.8.8",
+        var proxyApps: Boolean = false,
+        var bypass: Boolean = false,
+        var udpdns: Boolean = false,
+        var ipv6: Boolean = true,
+        var individual: String = "",
+        var tx: Long = 0,
+        var rx: Long = 0,
+        var userOrder: Long = 0,
+        var plugin: String? = null,
+        var udpFallback: Long? = null,
+
+        @Ignore // not persisted in db, only used by direct boot
+        var dirty: Boolean = false
+) : Parcelable, Serializable {
     companion object {
         private const val TAG = "ShadowParser"
         private const val serialVersionUID = 0L
@@ -199,29 +225,6 @@ class Profile : Serializable {
         @Query("DELETE FROM `Profile`")
         fun deleteAll(): Int
     }
-
-    @PrimaryKey(autoGenerate = true)
-    var id: Long = 0
-    var name: String? = ""
-    var host: String = "198.199.101.152"
-    var remotePort: Int = 8388
-    var password: String = "u1rRWTssNv0p"
-    var method: String = "aes-256-cfb"
-    var route: String = "all"
-    var remoteDns: String = "8.8.8.8"
-    var proxyApps: Boolean = false
-    var bypass: Boolean = false
-    var udpdns: Boolean = false
-    var ipv6: Boolean = true
-    var individual: String = ""
-    var tx: Long = 0
-    var rx: Long = 0
-    var userOrder: Long = 0
-    var plugin: String? = null
-    var udpFallback: Long? = null
-
-    @Ignore // not persisted in db, only used by direct boot
-    var dirty: Boolean = false
 
     val formattedAddress get() = (if (host.contains(":")) "[%s]:%d" else "%s:%d").format(host, remotePort)
     val formattedName get() = if (name.isNullOrEmpty()) formattedAddress else name!!