+++ mobile/src/main/java/com/github/shadowsocks/ProfileConfigActivity.kt
@@ -21,18 +21,30 @@
 package com.github.shadowsocks
 
 import android.app.Activity
+import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import com.github.shadowsocks.plugin.AlertDialogFragment
+import com.github.shadowsocks.plugin.Empty
 import com.github.shadowsocks.plugin.PluginContract
 import com.github.shadowsocks.preference.DataStore
 
 class ProfileConfigActivity : AppCompatActivity() {
     companion object {
         const val REQUEST_CODE_PLUGIN_HELP = 1
+    }
+
+    class UnsavedChangesDialogFragment : AlertDialogFragment<Empty, Empty>() {
+        override fun AlertDialog.Builder.prepare(listener: DialogInterface.OnClickListener) {
+            setTitle(R.string.unsaved_changes_prompt)
+            setPositiveButton(R.string.yes, listener)
+            setNegativeButton(R.string.no, listener)
+            setNeutralButton(android.R.string.cancel, null)
+        }
     }
 
     private val child by lazy { supportFragmentManager.findFragmentById(R.id.content) as ProfileConfigFragment }
@@ -59,13 +71,8 @@ class ProfileConfigActivity : AppCompatActivity() {
     override fun onOptionsItemSelected(item: MenuItem) = child.onOptionsItemSelected(item)
 
     override fun onBackPressed() {
-        if (DataStore.dirty) AlertDialog.Builder(this)
-                .setTitle(R.string.unsaved_changes_prompt)
-                .setPositiveButton(R.string.yes) { _, _ -> child.saveAndExit() }
-                .setNegativeButton(R.string.no) { _, _ -> finish() }
-                .setNeutralButton(android.R.string.cancel, null)
-                .create()
-                .show() else super.onBackPressed()
+        if (DataStore.dirty) UnsavedChangesDialogFragment().show(child, ProfileConfigFragment.REQUEST_UNSAVED_CHANGES)
+        else super.onBackPressed()
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {